+++ nextgisweb/webmap/nodepkg/print-map/PrintMap.tsx
@@ -2,13 +2,24 @@ import debounce from "lodash-es/debounce";
 import Map from "ol/Map";
 import type OlMap from "ol/Map";
 import View from "ol/View";
+import { Rotate } from "ol/control";
 import type { Coordinate } from "ol/coordinate";
 import { defaults as defaultInteractions } from "ol/interaction";
-import { useEffect, useRef, useState } from "react";
+import { useEffect, useMemo, useRef, useState } from "react";
+import { Rnd } from "react-rnd";
 
+import { gettext } from "@nextgisweb/pyramid/i18n";
+
+import { LayersTree } from "../layers-tree/LayersTree";
+import type { TreeWebmapItem } from "../layers-tree/LayersTree";
+import { getLabel } from "../map-controls/map-controls";
 import MapScaleControl from "../ol/ol-ext/ol-mapscale";
+import type WebmapStore from "../store";
 import type { DojoDisplay } from "../type";
+import type { LayerItem } from "../type/TreeItems";
 
+import { printMapStore } from "./PrintMapStore";
+import type { RndCoords } from "./PrintMapStore";
 import { buildPrintStyle } from "./PrintMapStyle";
 
 import "./PrintMap.less";
@@ -25,6 +36,26 @@ const setMapScale = (scale: number, olMap: OlMap): void => {
     const pointResolution3857 = cosh(center[1] / 6378137);
     const resolution = pointResolution3857 * (scale / (96 * 39.3701));
     olMap.getView().setResolution(resolution);
+};
+
+const switchRotateControl = (olMap: OlMap, show: boolean): void => {
+    const controls = olMap.getControls();
+    const rotateControl = controls
+        .getArray()
+        .find((control) => control instanceof Rotate);
+
+    if (!rotateControl && show) {
+        const rotateControl = new Rotate({
+            tipLabel: gettext("Reset rotation"),
+            label: getLabel("arrow_upward"),
+            autoHide: false,
+        });
+        olMap.addControl(rotateControl);
+    }
+
+    if (rotateControl && !show) {
+        olMap.removeControl(rotateControl);
+    }
 };
 
 const buildMap = (container: HTMLElement, display: DojoDisplay): OlMap => {
@@ -85,10 +116,13 @@ export interface PrintMapSettings {
     width: number;
     height: number;
     margin: number;
-    scale?: number;
     scaleLine: boolean;
     scaleValue: boolean;
+    arrow: boolean;
+    legend: boolean;
     center?: Coordinate | null;
+    title?: boolean;
+    scale?: number;
 }
 
 interface PrintMapProps {
@@ -138,6 +172,88 @@ class PrintMapStyle {
     }
 }
 
+const handleTreeItem = (checked: Set<number>, layersItem: TreeWebmapItem) => {
+    const { key } = layersItem;
+
+    if (layersItem.isLeaf) {
+        if (!checked.has(key)) {
+            return null;
+        }
+
+        const hasLegend =
+            layersItem.legendIcon ||
+            (layersItem.treeItem as LayerItem).legendInfo.open;
+
+        if (hasLegend) {
+            return layersItem;
+        } else {
+            return null;
+        }
+    }
+
+    if (layersItem.children) {
+        const newLayersItems: TreeWebmapItem[] = [];
+        (layersItem.children as TreeWebmapItem[]).forEach((item) => {
+            const newTreeItem = handleTreeItem(checked, item);
+            if (newTreeItem) {
+                newLayersItems.push(newTreeItem);
+            }
+        });
+        layersItem.children = newLayersItems;
+        return layersItem.children.length ? layersItem : null;
+    }
+};
+
+const filterTreeItems = (store: WebmapStore, layersItems: TreeWebmapItem[]) => {
+    const newLayersItems: TreeWebmapItem[] = [];
+    const checked = new Set(store.getChecked());
+    layersItems.forEach((item) => {
+        const newTreeItem = handleTreeItem(checked, item);
+        if (newTreeItem) {
+            newLayersItems.push(newTreeItem);
+        }
+    });
+    return newLayersItems;
+};
+
+interface RndCompProps {
+    init: RndCoords;
+    onChange: (coords: RndCoords) => void;
+    className?: string;
+    children: string | JSX.Element | JSX.Element[];
+}
+
+export const RndComp = ({
+    init,
+    onChange,
+    className,
+    children,
+}: RndCompProps) => {
+    return (
+        <Rnd
+            default={init}
+            onResizeStop={(e, direction, ref, delta, position) => {
+                onChange({
+                    width: ref.offsetWidth,
+                    height: ref.offsetHeight,
+                    ...position,
+                });
+            }}
+            onDragStop={(e, d) => {
+                onChange({
+                    width: printMapStore.legend.width,
+                    height: printMapStore.legend.height,
+                    x: d.x,
+                    y: d.y,
+                });
+            }}
+            className={className}
+        >
+            {children}
+        </Rnd>
+    );
+};
+
 export const PrintMap = ({
     settings,
     display,
@@ -145,11 +261,21 @@ export const PrintMap = ({
     onScaleChange,
     onCenterChange,
 }: PrintMapProps) => {
-    const { width, height, margin, scale, scaleLine, scaleValue } = settings;
+    const {
+        width,
+        height,
+        margin,
+        scale,
+        scaleLine,
+        scaleValue,
+        legend,
+        title,
+        arrow,
+    } = settings;
     const printMapRef = useRef<HTMLDivElement>(null);
     const printMap = useRef<OlMap>();
-    const [mapScaleControl, setMapScaleControl] = useState<MapScaleControl>();
 
+    const [mapScaleControl, setMapScaleControl] = useState<MapScaleControl>();
     const [style, setStyle] = useState<PrintMapStyle>();
 
     useEffect(() => {
@@ -213,6 +339,12 @@ export const PrintMap = ({
     }, [width, height, margin, scale, style]);
 
     useEffect(() => {
+        if (printMap.current) {
+            switchRotateControl(printMap.current, arrow);
+        }
+    }, [arrow]);
+
+    useEffect(() => {
         const control = mapScaleControl;
         if (control) {
             control.setScaleLineVisibility(scaleLine);
@@ -220,13 +352,68 @@ export const PrintMap = ({
         }
     }, [scaleLine, scaleValue, mapScaleControl]);
 
-    return (
-        <div className="print-map-page-wrapper">
-            <div className="print-map-export-wrapper">
-                <div className="print-map-page">
-                    <div className="print-map" ref={printMapRef}></div>
+    const legendComp = useMemo(() => {
+        if (!legend) {
+            return null;
+        }
+
+        return (
+            <RndComp
+                init={printMapStore.legend}
+                onChange={(rndCoords) => {
+                    printMapStore.legend = rndCoords;
+                }}
+                className="legend-rnd"
+            >
+                <div className="legend">
+                    <LayersTree
+                        {...{
+                            store: display.webmapStore,
+                            onSelect: () => {},
+                            setLayerZIndex: () => {},
+                            getWebmapPlugins: () => ({}),
+                            onReady: () => {},
+                            showDropdown: false,
+                            draggable: false,
+                            checkable: false,
+                            onFilterItems: (store, layersItems) => {
+                                return filterTreeItems(store, layersItems);
+                            },
+                        }}
+                    />
+                </div>
+            </RndComp>
+        );
+    }, [legend]);
+
+    const titleComp = useMemo(() => {
+        if (!title) return null;
+        return (
+            <RndComp
+                init={printMapStore.title}
+                onChange={(rndCoords) => {
+                    printMapStore.title = rndCoords;
+                }}
+                className="title-rnd"
+            >
+                <div className="title">{display.config.webmapTitle}</div>
+            </RndComp>
+        );
+    }, [title]);
+
+    if (printMap)
+        return (
+            <div className="print-map-page-wrapper">
+                <div className="print-map-export-wrapper">
+                    <div className="print-map-page">
+                        <div className="print-map">
+                            <div className="print-olmap" ref={printMapRef}>
+                                {legendComp}
+                                {titleComp}
+                            </div>
+                        </div>
+                    </div>
                 </div>
             </div>
-        </div>
-    );
+        );
 };